% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_snippets.R
\name{add_snippets}
\alias{add_snippets}
\title{Add the snippets to your ~/.R/snippets/r.snippets}
\usage{
add_snippets(rs = "~/.R/snippets/r.snippets")
}
\arguments{
\item{rs}{Path to R snippets}
}
\value{
A message about the success of the copy.
}
\description{
Add the snippets to your ~/.R/snippets/r.snippets
}
\examples{
\dontrun{
add_snippets()
}
}
